Infinite Stream recognition - Updates after review
