+++ speech/cloud-client/pom.xml
@@ -135,5 +135,33 @@
         </plugins>
       </build>
     </profile>
+    <profile>
+      <id>InfiniteStreamRecognize</id>
+      <activation>
+        <property>
+          <name>InfiniteStreamRecognize</name>
+        </property>
+      </activation>
+      <build>
+        <plugins>
+          <plugin>
+            <groupId>org.codehaus.mojo</groupId>
+            <artifactId>exec-maven-plugin</artifactId>
+            <version>1.6.0</version>
+            <executions>
+              <execution>
+                <goals>
+                  <goal>java</goal>
+                </goals>
+              </execution>
+            </executions>
+            <configuration>
+              <mainClass>com.example.speech.InfiniteStreamRecognize</mainClass>
+              <cleanupDaemonThreads>false</cleanupDaemonThreads>
+            </configuration>
+          </plugin>
+        </plugins>
+      </build>
+    </profile>
   </profiles>
 </project>
